requirements.txt
streamlit
tensorflow
numpy
pillow
opencv-python
protobuf==3.20.*
scikit-learn
matplotlib
